mac80211: ath11k: rework memory profile patch
ath-next now has a patch that changes QCN9074 FW mode and limits, so we
can use that to set the IPQ8074 FW mode and limits as well and not use
the current hacky way.

This also introduces a choice based menu to pick the memory profile as
there is potential for 256MB in future as well.

Signed-off-by: Robert Marko <robimarko@gmail.com>